feat: 모듈화 진행
