+++ pages/folder/folder.css
@@ -1,4 +1,4 @@
 .folder_introduction {
-  font-size: 5rem;
+  font-size: 3rem;
   color: blue;
 }
