+++ pages/sign/modules/message.js
@@ -1,0 +1,18 @@
+
+
+//에러 메세지가 담긴 객체
+
+const messages = {
+  email_empty_error : "이메일을 입력해 주세요",
+  email_type_error : "올바른 이메일 형식이 아닙니다",
+  email_check_error : "이메일을 확인해 주세요",
+  email_duplicate_error : "이미 사용중인 이메일입니다",
+  password_empty_error : "비밀번호를 입력해 주세요",
+  password_check_error: "비밀번호를 확인해 주세요",
+  password_type_error : "비밀번호는 영문, 숫자 조합 8자 이상 입력해 주세요",
+  password_repeat_error : "비밀번호가 일치하지 않아요"
+}
+
+
+//export
+export { messages }
