+++ pages/sign/modules/tags.js
@@ -1,0 +1,25 @@
+
+
+//이메일 관련 
+const email_input = document.querySelector('#email_input')
+const email_input_check = document.querySelector("#email_check")
+
+//패스워드 관련
+const password_input = document.querySelector('#password_input')
+const password_input_check = document.querySelector("#password_check")
+
+//패스워드 확인 관련
+const password_repeat_input = document.querySelector('#password_repeat_input')
+const password_repeat_input_check = document.querySelector('#password_repeat_check')
+
+//로그인, 회원가입 버튼
+const submit_button = document.querySelector('#submit_button')
+
+//눈 모양 아이콘 관련
+const eye_mark_in_password = document.querySelector('.eye_mark_in_password') 
+const eye_mark_in_password_repeat = document.querySelector('.eye_mark_in_password_repeat')
+
+
+//export
+export { email_input, email_input_check, password_input, password_input_check, password_repeat_input, password_repeat_input_check, submit_button, eye_mark_in_password, eye_mark_in_password_repeat}
+
