+++ pages/sign/signin.html
@@ -58,7 +58,7 @@
         </div>
       </div>          
     </div>
-    <script src="/pages/sign/signin.js"></script>
+    <script type="module" src="signin.js"></script>
     
   </main>
 </body>
