+++ pages/sign/signup.html
@@ -64,7 +64,7 @@
         </div>
       </div>          
     </div>
-    <script src="signup.js"></script>
+    <script type="module" src="signup.js"></script>
     
   </main>
 </body>
